Improved order handling of header cards, by separating additions, insertions and in-situ updates. Below is the differentiation of these:
 * Additions: Header.add***() methods, add new cards to the end of the header
   always, deleting prior occurences as needed. The current position remains
   unchanged by the additions.

 * Insertions: Header.insertLine() method inserts new cards at the current
   header position (e.g. as set by Header.findCard(), or prior insertions),
   and advancing the current position by one. Prior occurences of cards with
   the same key will be removed as necessary.

 * Updates: Header.updateLine() method updates an existing card (if any) that
   matches the 'key' (1st) argument, with a new card (2nd argument). The
   current position remains unaffected.

Part of the problem before was that the existence of prior entries under the
same key produced unpredictable header ordering. Now, the user can either add
new cards to the end of the header, just as add() would append entries to the
end of a List; or insert them at specific locations; or else update existing
keys in situ.