I10-format ASCII table columns should be long, not integer
Prior to this fix, ASCII table columns with format I10 were
interpreted as 32-bit integers (on the grounds that Integer.MAX_VALUE
has 10 characters?), but some 10-character numbers cannot be
represented in 32 bits, which meant that some potential values
in 10-character columns (in the range 2147483648-9999999999)
could not be read correctly.  Fix it so that I10 and larger,
not I11 and larger, are treated as 64-bit.