add ArrayDataInput.markSupported() method
This commit adds the method markSupported() to the ArrayDataInput
interface; it should really have been added at the same time as
the mark(int) and reset() methods were added, by analogy with the
InputStream class with reference to which they are defined.

Since all(?) known ArrayDataInput implementations currently support
mark/reset functionality this change should not be disruptive,
but it provides the option for future implementations to avoid
supporting mark/reset, which could be burdensome or impossible.

Application code using mark/reset ought to change, checking
markSupported returns true before using this functionality now.
But since all implementations currently return true, at the moment
it's not likely to break anything if they don't.